src/python/requirements.txt
libusb1==3.2.0
ndeflib==0.3.3
nfcpy==1.0.4
pyDes==2.0.1
pyserial==3.5
